+++ public/index.html
@@ -28,7 +28,6 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="login-modal"></div>
     <div id="root"></div>
     <!--
       This HTML file is a template.
