+++ src/components/Header.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { OPEN_LOGIN_MODAL } from '../store/reducers/user';
+import { OPEN_MODAL } from '../store/reducers/modal';
 
 const Header = () => {
   const dispatch = useDispatch();
@@ -19,7 +19,7 @@ const Header = () => {
               <Link to="/">서비스 소개</Link>
             </li>
             <li>
-              <Link to="/" onClick={() => dispatch({ type: OPEN_LOGIN_MODAL })}>
+              <Link to="/" onClick={() => dispatch({ type: OPEN_MODAL })}>
                 로그인
               </Link>
             </li>
