+++ src/index.jsx
@@ -16,13 +16,6 @@ ReactDOM.render(
   document.getElementById('root'),
 );
 
-const ModalContainer = ({ children }) => {
-  const el = document.getElementById('login-modal');
-  return ReactDOM.createPortal(children, el);
-};
-
-export default ModalContainer;
-
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
