+++ src/store/reducers/index.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux';
 import team from './team';
-import user from './user';
+import modal from './modal';
 
 // 리듀서들을 하나로 합침
 export default combineReducers({
   team,
-  user,
+  modal,
 });
