+++ src/store/reducers/modal.js
@@ -1,0 +1,32 @@
+import produce from 'immer';
+
+export const OPEN_MODAL = 'modal/OPEN__MODAL';
+export const CLOSE_MODAL = 'modal/CLOSE_MODAL';
+
+export const OpenModal = () => ({
+  type: OPEN_MODAL,
+});
+export const CloseModal = () => ({
+  type: CLOSE_MODAL,
+});
+
+const initialState = {
+  openYn: false,
+};
+
+const reducer = (state = initialState, action) => {
+  return produce(state, draft => {
+    switch (action.type) {
+      case OPEN_MODAL: {
+        draft.openYn = true;
+        return draft;
+      }
+      case CLOSE_MODAL: {
+        draft.openYn = false;
+        return draft;
+      }
+    }
+  });
+};
+
+export default reducer;
