+++ src/styles/components/_LoginForm.scss
@@ -1,0 +1,68 @@
+.login-header {
+  font-size: 0.5rem;
+  display: flex;
+  flex-direction: row;
+
+  .point {
+    font-size: 1rem;
+  }
+}
+.main-logo {
+  text-align: center;
+  width: 100%;
+}
+.close-box {
+  margin-left: auto;
+  font-size: 0.9rem;
+  color: #bbbbbb;
+}
+
+.login-contents {
+  padding-top: 5%;
+  text-align: center;
+  .login-title {
+    font-size: 1.5rem;
+    margin: 0;
+  }
+  .login-content {
+    padding-top: 3%;
+    font-size: 0.88rem;
+    color: #777777;
+    padding-bottom: 5%;
+  }
+}
+.line {
+  border: 1px solid #eeeeee;
+  margin-top: 4% 0;
+}
+.email-input {
+  margin-bottom: 5%;
+  height: 3.4rem;
+}
+.email-button {
+  color: #ffffff;
+  height: 3.4rem;
+  background-color: $main-color-blue;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+.github-button {
+  color: #272e40;
+  height: 3.4rem;
+  display: flex;
+  background-color: #adb5bd;
+  justify-content: space-between;
+  align-items: center;
+}
+.email-icon,
+.github-icon {
+  font-size: 1.2rem;
+  justify-content: flex-start;
+  svg {
+    margin-top: 15%;
+  }
+}
+#right-space {
+  opacity: 0;
+}
