+++ src/styles/components/_Modal.scss
@@ -4,73 +4,46 @@
   position: fixed;
   left: 0;
   top: 0;
-  height: 100%;
-  width: 100%;
+  height: 100vh;
+  width: 100vw;
   display: flex;
   align-items: center;
-  flex-direction: row;
+  flex-direction: column;
   justify-content: center;
 }
 
-.modal-container .content {
+.modal-container .modal-content {
   background: white;
   padding: 1rem;
-  width: 390px;
   height: auto;
-  border-radius: 5px 5px 5px 5px;
-}
-.line {
-  border: thin solid #eeeeee;
-  margin-top: 4%;
-  margin-bottom: 4%;
-}
-.main-logo {
-  font-size: 0.5rem;
-  text-align: center;
-}
-.right {
-  float: right;
-  font-size: 0.9rem;
-  color: #bbbbbb;
-}
-.point {
-  font-size: 1rem;
+  border-radius: 5px;
 }
 
-.login-contents {
-  padding-top: 5%;
-  text-align: center;
-  .login-title {
-    font-size: 1.5rem;
-    margin: 0;
-  }
-  .login-content {
-    padding-top: 3%;
-    font-size: 0.88rem;
-    color: #777777;
-    padding-bottom: 5%;
-  }
+// modal size
+.modal-xs {
+  width: 500px;
 }
-.email-input {
-  margin-bottom: 5%;
-  height: 3.4rem;
-}
-.email-button {
-  color: #ffffff;
-  height: 3.4rem;
-  background-color: #5f76f3;
-}
-.github-button {
-  color: #272e40;
-  height: 3.4rem;
-  background-color: #adb5bd;
-}
-.email-icon,
-.github-icon {
-  float: left;
-  font-size: 1.2rem;
 
-  svg {
-    margin-top: 15%;
-  }
+.modal-sm {
+  width: 600px;
+}
+
+.modal-md {
+  width: 700px;
+}
+
+.modal-lg {
+  width: 800px;
+}
+
+.modal-xl {
+  width: 900px;
+}
+
+.modal-login {
+  width: 390px;
+}
+
+.modal-full {
+  width: 95%;
 }
