+++ src/styles/main.scss
@@ -173,3 +173,4 @@ button {
 @import './components/_Header.scss';
 @import './components/_Banner.scss';
 @import './components/Modal';
+@import './components/LoginForm';
